+++ R/vsearch.R
@@ -164,9 +164,7 @@ vs_search_global <- function(physeq,
 #'   \url{https://github.com/torognes/swarm/}.
 #'
 #' @export
-#' @examples
-#' \dontrun{
-#'
+#' @examplesIf MiscMetabar:::is_swarm_installed()
 #' summary_plot_pq(data_fungi)
 #' system2("swarm", "-h")
 #'
@@ -188,13 +186,12 @@ vs_search_global <- function(physeq,
 #' sequences_ex_swarm <- swarm_clustering(
 #'   dna_seq = sequences_ex
 #' )
-#' }
 #' @seealso [asv2otu()], [vsearch_clustering()]
 #' @references
 #'   SWARM can be downloaded from
 #'   \url{https://github.com/torognes/swarm}.
 #'   More information in the associated publications
-#'   \doi{doi:10.1093/bioinformatics/btab493} and \doi{doi:10.7717/2Fpeerj.593}
+#'   \doi{doi:10.1093/bioinformatics/btab493} and \doi{doi:10.7717/peerj.593}
 #' @details
 #' This function is mainly a wrapper of the work of others.
 #'   Please cite [SWARM](https://github.com/torognes/swarm).
@@ -348,7 +345,7 @@ swarm_clustering <- function(physeq = NULL,
 #'   set tax_adjust to 0 (default). For the moment only tax_adjust = 0 is
 #'   robust
 #' @param vsearch_cluster_method (default: "--cluster_size) See other possible
-#'   methods in the [vsearch pdf manual](https://github.com/torognes/vsearch/releases/download/v2.23.0/vsearch_manual.pdf) (e.g. `--cluster_size` or `--cluster_smallmem`)
+#'   methods in the [vsearch manual](https://github.com/torognes/vsearch) (e.g. `--cluster_size` or `--cluster_smallmem`)
 #'   - `--cluster_fast` : Clusterize the fasta sequences in filename, automatically sort by decreasing sequence length beforehand.
 #'   - `--cluster_size` : Clusterize the fasta sequences in filename, automatically sort by decreasing sequence abundance beforehand.
 #'   - `--cluster_smallmem` : Clusterize the fasta sequences in filename without automatically modifying their order beforehand. Sequence are expected to be sorted by decreasing sequence length, unless *--usersort* is used.
@@ -376,8 +373,7 @@ swarm_clustering <- function(physeq = NULL,
 #' @export
 #' @author Adrien Taudière
 #'
-#' @examples
-#'
+#' @examplesIf MiscMetabar:::is_vsearch_installed()
 #' summary_plot_pq(data_fungi)
 #' d_vs <- vsearch_clustering(data_fungi)
 #' summary_plot_pq(d_vs)
@@ -503,8 +499,7 @@ vsearch_clustering <- function(physeq = NULL,
 #'
 #' @export
 #'
-#' @examples
-#' \dontrun{
+#' @examplesIf MiscMetabar:::is_vsearch_installed()
 #' data_fungi_nochim <- chimera_removal_vs(data_fungi)
 #' data_fungi_nochim_16 <- chimera_removal_vs(data_fungi,
 #'   abskew = 16,
@@ -514,7 +509,6 @@ vsearch_clustering <- function(physeq = NULL,
 #'   chimera_removal_vs(data_fungi, type = "Select_only_non_chim")
 #' data_fungi_chimera <-
 #'   chimera_removal_vs(data_fungi, type = "Select_only_chim")
-#' }
 #' @author Adrien Taudière
 #' @details
 #' This function is mainly a wrapper of the work of others.
@@ -626,13 +620,11 @@ chimera_removal_vs <-
 #'   (`$chimera`) and bordeline taxa (`$borderline`)
 #' @export
 #'
-#' @examples
-#' \dontrun{
+#' @examplesIf MiscMetabar:::is_vsearch_installed()
 #' chimera_detection_vs(
 #'   seq2search = data_fungi@refseq,
 #'   nb_seq = taxa_sums(data_fungi)
 #' )
-#' }
 #' @author Adrien Taudière
 #' @details
 #' This function is mainly a wrapper of the work of others.