+++ NEWS.md
@@ -6,6 +6,10 @@
 - Add functions `swarm_clustering()` and `vsearch_clustering()` and add `swarm` method in the function `asv2otu()`
 - Add function `physeq_or_string_to_dna()` mostly for internal use
 - Add function `cutadapt_remove_primers()` to remove primers using [cutadapt](https://github.com/marcelm/cutadapt/)
+- Add internal functions `is_swarm_installed()`, `is_cutadapt_installed()`, `is_vsearch_installed()` and `is_falco_installed()` to test for the availability of external software in order to run examples and test from testthat.
+
+- Submit to CRAN and change code to comply with their rules (patch 0.7.1 to 0.7.4)
+
 
 ## BREAKING CHANGES